+++ packages/assignment-5/src/refactoring/hooks/useCartCalculator.ts
@@ -1,0 +1,9 @@
+import { calculateCartTotal } from './utils/cartUtils';
+
+export const useCartCalculator = (cart, selectedCoupon) => {
+  const calculateTotal = () => {
+    return calculateCartTotal(cart, selectedCoupon);
+  };
+
+  return { calculateTotal };
+};
